Check year-pos for Roman-numerals
This avoids titles with `i` in the middle of the title (directly translated as "in", for danish titles) being converted to roman numerals (capital I, II, III)

Example
```
Nattevagten - Daemoner gaar i arv
```
Would wrongly be interpreted as roman numerals and replace with a number instead of a letter, which in this case means "in" directly translated.
```
Nattevagten - Daemoner gaar 1 arv
```

This commit assumes the roman numerals are directly before the year-tag. Personally I've never witnessed movies with roman numerals in the middle of the title.

If you think this needs to be broken up in several functions I agree, but since parse() is one big function I'll let you decide how to proceed.